refac: split admin page into multiple tab files
