+++ playwright-tests/tests/admin.spec.js
@@ -6,103 +6,184 @@ test.describe("Wallet is connected", () => {
     storageState: "playwright-tests/storage-states/wallet-connected-admin.json",
   });
 
-  test("should show be able to fill in the form to create a new team", async ({
+  test("should be able to manage featured communities from home page settings tab", async ({
     page,
   }) => {
     await page.goto("/devhub.near/widget/app?page=admin");
 
-    const buttonSelector = `button[data-testid="create-team"]`;
+    const buttonSelector = `button[data-testid="preview-homepage"]`;
     // Wait for the first post history button to be visible
     await page.waitForSelector(buttonSelector, {
       state: "visible",
     });
 
     // Click on the first post history button
-    await page.click(buttonSelector);
-    await page.locator(".flex-grow-1").first().click();
-    await page.getByPlaceholder("Team name").click();
-    await page.getByPlaceholder("Team name").fill("new team");
-    await page.getByPlaceholder("Team name").press("Tab");
+    await page.getByPlaceholder("Community handle").nth(4).click();
     await page
-      .frameLocator("iframe")
-      .getByRole("textbox")
-      .fill("new team description");
-    await page.getByLabel("Select type").selectOption("starts-with:");
-    await page.locator("#editPostCheckbox").check();
-    await page.locator("#editPostCheckbox").uncheck();
-    await page.locator("#useLabelsCheckbox").check();
-    await page.locator("#useLabelsCheckbox").uncheck();
-    await page.getByPlaceholder("label").click();
-    await page.getByPlaceholder("label").fill("restrictivelabel");
-    await page.getByPlaceholder("member").click();
-    await page.getByPlaceholder("member").fill("thomasguntenaar.near");
-    const addMemberButton = await page.waitForSelector(
-      "button.btn.btn-success.add-member"
-    );
-    await addMemberButton.click();
-    await page.getByPlaceholder("member").nth(1).click();
-    await page.getByPlaceholder("member").nth(1).fill("theori.near");
-    addMemberButton.click();
-
-    await page.getByRole("button", { name: "Submit" }).click();
-  });
-
-  test("should hide team configurator when hit cancel", async ({ page }) => {
-    await page.goto("/devhub.near/widget/app?page=admin");
-    const buttonSelector = `button[data-testid="create-team"]`;
-    // Wait for the first post history button to be visible
-    await page.waitForSelector(buttonSelector, {
-      state: "visible",
-    });
-
-    await page.getByRole("button", { name: " Edit team" }).first().click();
-    await page.locator("div:nth-child(10) > .btn").click();
-    await page.locator("div:nth-child(9) > .btn").click();
-    await page.getByRole("button", { name: "" }).nth(4).click();
+      .getByPlaceholder("Community handle")
+      .nth(4)
+      .fill("thomasguntenaar");
+    await page.getByTestId("add-to-list").click();
+    await page.getByRole("button", { name: " Submit" }).click();
+    await page.getByText("Close").click();
     await page.getByRole("button", { name: "" }).nth(3).click();
-    await page.getByRole("button", { name: "" }).nth(2).click();
-    await page.getByRole("button", { name: "" }).nth(1).click();
+    await page.getByRole("button", { name: " Submit" }).click();
+    await page.getByText("Can't set fewer than 4 communities").click();
+    await page
+      .getByText(
+        "NEAR BOS embeddable custom element /admin /communities /activity feed /about ↓ H"
+      )
+      .press("Escape");
     await page.getByRole("button", { name: "Cancel" }).click();
+
+    await page.getByTestId("preview-homepage").click();
+    await page.getByRole("heading", { name: "/connect" }).click();
+    await page.getByTestId("preview-homepage").click();
   });
 
-  test("should be able to toggle the manage featured communities section", async ({
-    page,
-  }) => {
+  test("should be able to manage moderators", async ({ page }) => {
     await page.goto("/devhub.near/widget/app?page=admin");
-    const buttonSelector = `button[data-testid="manage-featured"]`;
-    // Wait for the first post history button to be visible
+    const buttonSelector = `button[data-testid="preview-homepage"]`;
+    // Wait for the first post button to be visible
     await page.waitForSelector(buttonSelector, {
       state: "visible",
     });
 
-    // This part removes 2 of the featured communities and clicks cancel
-    // They should appear again
-    await page.getByTestId("manage-featured").click();
-    await page.getByRole("button", { name: "Cancel" }).first().click();
-    await page.getByTestId("manage-featured").click();
-    await page.locator("div:nth-child(5) > .btn").click();
-    await page.getByRole("button", { name: "" }).nth(2).click();
-    await page.getByRole("button", { name: "Cancel" }).first().click();
-    await page.getByTestId("manage-featured").click();
-    await page.locator("div:nth-child(5) > .btn").click();
-    await page.getByRole("button", { name: "Cancel" }).first().click();
+    await page.getByRole("tab", { name: "Moderators" }).click();
+    await page.getByTestId("edit-members").click();
+    await page.locator("div:nth-child(9)").click();
+    await page
+      .locator(
+        "div:nth-child(9) > div > .d-flex > .input-group > .form-control"
+      )
+      .click();
+    await page
+      .locator(
+        "div:nth-child(9) > div > .d-flex > .input-group > .form-control"
+      )
+      .fill("thomasguntenaar.near");
+    await page.getByRole("button", { name: "" }).click();
+    await page.getByRole("button", { name: " Submit" }).click();
+    await page.getByText("Close").click();
+    await page.getByRole("button", { name: "Cancel" }).click();
+    await page.getByTestId("edit-members").click();
+    await page.locator("div:nth-child(9)").click();
+    await page.getByTestId("edit-members").click();
+  });
+
+  test("should be able to manage restricted labels", async ({ page }) => {
+    await page.goto("/devhub.near/widget/app?page=admin");
+    const buttonSelector = `button[data-testid="preview-homepage"]`;
+    // Wait for the first post button to be visible
+    await page.waitForSelector(buttonSelector, {
+      state: "visible",
+    });
+    await page.getByRole("tab", { name: "Restricted labels" }).click();
+    await page.getByTestId("create-team").click();
+    await page.getByRole("button", { name: "Cancel" }).click();
+    await page.getByTestId("create-team").click();
+    await page
+      .getByLabel("This team is allowed to edit-post with this / these labels")
+      .check();
+    await page
+      .getByLabel("Only this team and moderators are allowed to use this label")
+      .check();
+    await page.getByPlaceholder("Team name").click();
+    await page.getByPlaceholder("Team name").fill("tom");
+    await page.getByPlaceholder("Team name").press("Tab");
+    await page.getByLabel("Select type").press("Tab");
+    await page.getByPlaceholder("label").fill("thom");
+    await page.getByPlaceholder("label").press("Tab");
+    await page
+      .getByLabel("This team is allowed to edit-post with this / these labels")
+      .press("Tab");
+    await page
+      .getByLabel("Only this team and moderators are allowed to use this label")
+      .press("Tab");
+    await page.getByPlaceholder("member").fill("tom.near");
+    await page.getByRole("button", { name: "" }).click();
+    await page.getByRole("button", { name: " Submit" }).click();
+    await page.getByLabel("Close").click();
+    await page
+      .getByRole("row", {
+        name: "mnw Multiple labels with common prefix No members in this group  Edit",
+      })
+      .getByRole("button")
+      .click();
+    await page
+      .getByRole("rowheader", {
+        name: "Create label Group name Team name Would you like this group to limit their restrictions to a single label, or would you prefer them to restrict it with any label that follows a similar convention? Select type What would you like the restricted label to be? starts-with: label Select label permissions This team is allowed to edit-post with this / these labels Only this team and moderators are allowed to use this label member member  Cancel  Submit",
+      })
+      .getByPlaceholder("member")
+      .click();
+    await page
+      .getByRole("rowheader", {
+        name: "Create label Group name Team name Would you like this group to limit their restrictions to a single label, or would you prefer them to restrict it with any label that follows a similar convention? Select type What would you like the restricted label to be? starts-with: label Select label permissions This team is allowed to edit-post with this / these labels Only this team and moderators are allowed to use this label member member  Cancel  Submit",
+      })
+      .getByPlaceholder("member")
+      .fill("thomas.near");
+    await page.getByRole("table").getByRole("button", { name: "" }).click();
+    await page
+      .getByRole("rowheader", {
+        name: "Create label Group name Team name Would you like this group to limit their restrictions to a single label, or would you prefer them to restrict it with any label that follows a similar convention? Select type What would you like the restricted label to be? starts-with: label Select label permissions This team is allowed to edit-post with this / these labels Only this team and moderators are allowed to use this label member member  member member  Cancel  Submit",
+      })
+      .getByLabel("This team is allowed to edit-post with this / these labels")
+      .uncheck();
+    await page
+      .getByRole("rowheader", {
+        name: "Create label Group name Team name Would you like this group to limit their restrictions to a single label, or would you prefer them to restrict it with any label that follows a similar convention? Select type What would you like the restricted label to be? starts-with: label Select label permissions This team is allowed to edit-post with this / these labels Only this team and moderators are allowed to use this label member member  member member  Cancel  Submit",
+      })
+      .getByLabel("This team is allowed to edit-post with this / these labels")
+      .uncheck();
+    await page
+      .getByRole("rowheader", {
+        name: "Create label Group name Team name Would you like this group to limit their restrictions to a single label, or would you prefer them to restrict it with any label that follows a similar convention? Select type What would you like the restricted label to be? starts-with: label Select label permissions This team is allowed to edit-post with this / these labels Only this team and moderators are allowed to use this label member member  member member  Cancel  Submit",
+      })
+      .getByLabel("Only this team and moderators are allowed to use this label")
+      .uncheck();
+    await page
+      .getByRole("rowheader", {
+        name: "Create label Group name Team name Would you like this group to limit their restrictions to a single label, or would you prefer them to restrict it with any label that follows a similar convention? Select type What would you like the restricted label to be? starts-with: label Select label permissions This team is allowed to edit-post with this / these labels Only this team and moderators are allowed to use this label member member  member member  Cancel  Submit",
+      })
+      .getByPlaceholder("label")
+      .click();
+    await page
+      .getByRole("rowheader", {
+        name: "Create label Group name Team name Would you like this group to limit their restrictions to a single label, or would you prefer them to restrict it with any label that follows a similar convention? Select type What would you like the restricted label to be? starts-with: label Select label permissions This team is allowed to edit-post with this / these labels Only this team and moderators are allowed to use this label member member  member member  Cancel  Submit",
+      })
+      .getByPlaceholder("Team name")
+      .click({
+        clickCount: 3,
+      });
+    await page
+      .getByRole("rowheader", {
+        name: "Create label Group name Team name Would you like this group to limit their restrictions to a single label, or would you prefer them to restrict it with any label that follows a similar convention? Select type What would you like the restricted label to be? starts-with: label Select label permissions This team is allowed to edit-post with this / these labels Only this team and moderators are allowed to use this label member member  member member  Cancel  Submit",
+      })
+      .getByPlaceholder("Team name")
+      .fill("new-group-name");
+    await page
+      .getByRole("table")
+      .getByRole("button", { name: " Submit" })
+      .click();
+    await page.getByLabel("Close").click();
   });
 
   test("shouldn't be able to add a none existing community handle without a warning", async ({
     page,
   }) => {
     await page.goto("/devhub.near/widget/app?page=admin");
-    const buttonSelector = `button[data-testid="manage-featured"]`;
-    // Wait for the first post history button to be visible
+    const buttonSelector = `button[data-testid="preview-homepage"]`;
+    // Wait for the first post  button to be visible
     await page.waitForSelector(buttonSelector, {
       state: "visible",
     });
-    await page.getByTestId("manage-featured").click();
-    await page.getByPlaceholder("zero-knowledge").click();
+    await page.getByPlaceholder("Community handle").nth(4).click();
     await page
-      .getByPlaceholder("zero-knowledge")
+      .getByPlaceholder("Community handle")
+      .nth(4)
       .fill("arandomnonsensehandlethatwouldnotexist");
-    await page.getByRole("button", { name: "" }).click();
+    await page.getByTestId("add-to-list").click();
+    await page.getByTestId("add-to-list").click();
     await page
       .getByText(
         "This community handle does not exist, make sure you use an existing handle."
