+++ src/app.jsx
@@ -165,7 +165,7 @@ function Page() {
     case "admin": {
       return (
         <Widget
-          src={"${REPL_DEVHUB}/widget/devhub.page.admin"}
+          src={"${REPL_DEVHUB}/widget/devhub.page.admin.index"}
           props={passProps}
         />
       );
